[FE] refactor: roi의 뷰포트 좌표 -> 이미지 좌표로 변환하는 hook 생성
